Add test_stack io tests
